move "each" files into detail folder since they cannot be used standalone
